implement w3c and b3 context propagators
This commit implements new http propagators in other to be compatible with
OpenTelemetry instrumented apps:
- w3c: default context propagator in opentelemetry
- b3: another choice
tracer.configure() was extended to accept a new parameter, "http_propagator",
this is a factory function that will be used to get new instances of the
propagator to be used.

The following is an example of how to set b3 as the propagator:

...
from oteltrace.propagation.b3 import B3HTTPPropagator

oteltrace.tracer.configure(
    http_propagator=B3HTTPPropagator
)
...

Users of oteltrace-run can choose the propagator to be used by setting the
OTEL_TRACER_PROPAGATOR env variable to one of the  current supported values:
w3c, b3 and datadog.