propagators: add API to set propagator type
6ba49ad7070c ("implement w3c and b3 context propagators") introduced support
for w3c and b3 propagators.  The configuration mechanism implemented in that
commit was based only on env variables.  This commit extends this behaviour
by adding a new "http_propagator" parameter to tracer.configure(), this
parameter allows to set the factory to be used to get new instances of the
propagator.

The following is an example of how to set b3 as the propagator:

...
from ddtrace.propagation.b3 import B3HTTPPropagator

ddtrace.tracer.configure(
    http_propagator=B3HTTPPropagator
)
...